Refactor configuration package layout
