+++ tests/unit/test_config_env.py
@@ -1,0 +1,41 @@
+from __future__ import annotations
+
+import pytest
+
+from library.config.env import (
+    _apply_env_overrides,
+    _expand_config_placeholders,
+    _resolve_placeholder_base_path,
+)
+
+
+@pytest.mark.unit
+def test_apply_env_overrides__alias(monkeypatch):
+    monkeypatch.setenv("CHEMBL_DA_LOG_LEVEL", "DEBUG")
+    data = {"system": {"log": {"level": "INFO"}}}
+
+    overrides = _apply_env_overrides(data)
+
+    assert data["system"]["log"]["level"] == "DEBUG"
+    assert overrides[("system", "log", "level")] == "CHEMBL_DA_LOG_LEVEL"
+
+
+@pytest.mark.unit
+def test_resolve_placeholder_base_path__env_override(monkeypatch, tmp_path):
+    base = tmp_path / "chembl"
+    monkeypatch.setenv("CHEMBL_DA_BASE_PATH", str(base))
+
+    resolved = _resolve_placeholder_base_path(None)
+
+    assert resolved == base.resolve()
+
+
+@pytest.mark.unit
+def test_expand_config_placeholders__replaces_marker(tmp_path):
+    base = tmp_path / "chembl"
+    base.mkdir()
+    data = {"paths": {"output": "$CHEMBL_DA_BASE_PATH/results"}}
+
+    expanded = _expand_config_placeholders(data, base_path=base)
+
+    assert expanded["paths"]["output"] == str(base / "results")
